consolidated CacheWhole, ResolverCache, and SplacheCache. Also changed package.json main and added express as a dependency
