--- src/SplacheCache.ts
@@ -1,158 +1,0 @@
-import { createClient, RedisClientType } from 'redis';
-import { parse, visit, GraphQLSchema } from 'graphql';
-import { graphql } from 'graphql';
-
-/* Creates an importable SplacheCache Class that accepts a GraphQL schema 
-and connects to the user's local Redis client or provided Redis client*/
-export class SplacheCache {
-  schema: GraphQLSchema;
-  typeToFields: object;
-  queryToReturnType: object;
-  client: RedisClientType;
-  constructor(
-    schema: GraphQLSchema,
-    host?: string,
-    port?: number,
-    password?: string
-  ) {
-    this.schema = schema;
-    this.GQLquery = this.GQLquery.bind(this);
-    if (host && port && password) {
-      this.client = createClient({
-        socket: {
-          host,
-          port,
-        },
-        password,
-      });
-    } else if (host && port) {
-      this.client = createClient({
-        socket: {
-          host,
-          port,
-        },
-      });
-    } else {
-      this.client = createClient();
-    }
-    this.client.connect()
-    .then(() => console.log('connected to redis'))
-    .catch((err) => console.log(`there was a problem connecting to the redis instance: ${err}`))
-  }
-  //Partial normalization of the input query through traversing the GraphQL AST to build valid root queries. 
-  //The root queries are checked against the Redis cache as opposed to the original query
-  async GQLquery(req: any, res: any, next: any) {
-    const queryString: string = req.body.query;
-    const ast = parse(queryString);
-    const [template, fieldArgs] = await makeTemplate(ast);
-    const splitQuery = qlStrGen(template, fieldArgs);
-    const compiledObj = { data: {} };
-    for (const query of splitQuery) {
-      const isInCache = await this.client.EXISTS(query);
-      if (isInCache) {
-        const returnObj = await this.client.GET(query);
-        if (typeof returnObj === 'string') {
-          const returnObjParsed = JSON.parse(returnObj);
-          compiledObj.data[Object.keys(returnObjParsed.data)[0]] =
-            returnObjParsed.data[Object.keys(returnObjParsed.data)[0]];
-        } else {
-          return next({ err: 'There was a redis error' });
-        }
-      } else {
-        await graphql({ schema: this.schema, source: query })
-          .then((response: any) => {
-            compiledObj.data[Object.keys(response.data)[0]] =
-              response.data[Object.keys(response.data)[0]];
-            this.client.SET(query, JSON.stringify(response));
-          })
-          .catch((err) => next({ err }));
-      }
-    }
-    res.locals.queryResult = compiledObj;
-    return next();
-  }
-}
-
-// makeTemplate uses the AST from the GraphQL query as an input
-// The visit function is provided by graphql, check documentation here https://graphql.org/graphql-js/language/#visit
-export async function makeTemplate(ast: any) {
-  const template: any = {};
-  const path: any = [];
-  const fieldInfo: any = {};
-  visit(ast, {
-    Field: {
-      enter(node: any) {
-        if (node.arguments) {
-          const args: any = {};
-          for (let i = 0; i < node.arguments.length; i++) {
-            const key = node.arguments[i].name.value;
-            args[key] = node.arguments[i].value.value;
-          }
-          fieldInfo[node.name.value] = { ...args };
-        }
-
-        path.push(node.name.value);
-      },
-      leave() {
-        path.pop();
-      },
-    },
-    SelectionSet: {
-      enter(node: any, key, parent: any) {
-        if (parent.kind === 'Field') {
-          const fields = {};
-          for (let i = 0; i < node.selections.length; i++) {
-            if (!node.selections[i].selectionSet) {
-              fields[node.selections[i].name.value] = true;
-            }
-          }
-          const fieldsObj = {
-            ...fields,
-            ...fieldInfo[path[path.length - 1]],
-          };
-          let curr = template;
-          for (let i = 0; i < path.length; i++) {
-            if (i + 1 === path.length) curr[path[i]] = { ...fieldsObj };
-            curr = curr[path[i]];
-          }
-        }
-      },
-      leave() {
-        path.pop();
-      },
-    },
-  });
-  return [template, fieldInfo];
-}
-
-//GQLquery helper functions below
-export function qlStrGen(template, fieldArgs) {
-  const queryStrs: string[] = [];
-  for (const prop in template) {
-    let queryStr = `query {${prop} ${genArgStr(fieldArgs[prop])} {`;
-    queryStr += genfields(template[prop]);
-    queryStrs.push((queryStr += '} }'));
-  }
-  return queryStrs;
-}
-
-export function genArgStr(args) {
-  if (Object.keys(args)[0] === undefined) return '';
-  let argStr = '(';
-  for (const arg in args) {
-    argStr += `${arg}: "${args[arg]}" `;
-  }
-  return (argStr += ')');
-}
-
-export function genfields(fields) {
-  let fieldStr = '';
-  for (const field in fields) {
-    if (typeof fields[field] === 'object') {
-      fieldStr += `${field} {${genfields(fields[field])}} `;
-    } else {
-      if (field !== 'id') fieldStr += `${field} `;
-    }
-  }
-  return fieldStr;
-}
